Added dictionary_builder file
This file defines two classes, Trie and WordDictionary, the first of which is used in the second to create a word dictionary into which words can be added via the addWord function. Following the addition of words, it can be checked if one is in the dictionary via the search function, which returns a bool True or False.